logic to check if the ingredient is in use before exclusion added

+++ src/main/java/br/com/flavorCRUD/service/IngredientServiceImpl.java
@@ -6,13 +6,18 @@ import java.util.List;
 import javax.inject.Inject;
 
 import br.com.flavorCRUD.dao.IngredientDao;
+import br.com.flavorCRUD.domain.Flavor;
 import br.com.flavorCRUD.domain.Ingredient;
+import br.com.flavorCRUD.util.FacesMessageUtil;
 
 public class IngredientServiceImpl implements IngredientService, Serializable{
 
 	@Inject
 	private IngredientDao dao;
 
+	@Inject
+	private FlavorService service;
+	
 	public void create(Ingredient ingredient) {
 		this.dao.create(ingredient);
 	}
@@ -22,10 +27,26 @@ public class IngredientServiceImpl implements IngredientService, Serializable{
 	}
 
 	public void delete(Ingredient ingredient) {
-		this.dao.delete(ingredient);
+		if(this.ingredientIsInAFalvor(ingredient)) {
+			new FacesMessageUtil().errorMessage("O ingrediente esta em uso em algum sabor");
+		} else {
+			this.dao.delete(ingredient);
+			new FacesMessageUtil().successMessage("Ingrediente removido com sucesso");
+		}
 	}
 
 	public void update(Ingredient ingredient) {
 		this.dao.update(ingredient);
+	}
+	
+	private boolean ingredientIsInAFalvor(Ingredient ingredient) {
+		boolean aux = false;
+		
+		for(Flavor f: this.service.getAll()) {
+			if(!aux)
+				aux = f.getIngredients().contains(ingredient);
+		}
+
+		return aux;
 	}
 }
